unit test for interest and fix a bug of 'int32/uint32'

+++ libraries/chain/asset_object.cpp
@@ -64,7 +64,7 @@ Interest asset_lock_data_object::_get_interest(uint32_t lock_period,const databa
    
    Interest		top_of_interest = fast_pow_of_interest(nominal_interest_perday, max_period);
    
-   int32_t pecent_of_year = (lock_days - max_period / 2)*GRAPHENE_100_PERCENT / (max_period / 2);
+   int32_t pecent_of_year = int32_t(lock_days - int32_t(max_period) / 2)*GRAPHENE_100_PERCENT / (int32_t(max_period) / 2);
 
    int64_t reward_rate = GRAPHENE_100_PERCENT + pecent_of_year * reward_coefficient / GRAPHENE_100_PERCENT;
